ci: separate formal and multi-arch-test-build
Separate formal and multi-arch-test-build so formal trigger can be
switched to pull_request_target. This event allows the workflow to do
things like label or comment on pull requests from forks, but it's not
recommended for build jobs due to security implications.

Label PRs in formal and build them multi-arch-test-build, if 'not
following guidelines' label is not set.

Link: https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#pull_request_target
Signed-off-by: George Sapkin <george@sapk.in>