ci: duplicate formal to fix posting comments
Duplicate formal so the event trigger can be set to pull_request_target
and fix posting comments. This event allows the workflow to do things
like label or comment on pull requests from forks, but it's not
recommended for build jobs due to security implications. The build jobs
still depends on formal that's why it has a duplicate one with less
privileges.

It doesn't look like there's an easy and secure way to have a workflow
with lower privileges (e.g. build) depend on workflow with higher ones
(e.g. formal or labeler that modify a PR). There's only the reverse with
going from lower privileges to higher ones with workflow_run, for
example when posting build results after a build to a PR.

Either switching existing combined workflow to pull_request_target or
splitting it into formal and build and switching build to workflow_run
gives build unsafe privileges. Splitting and switching build to
workflow_dispatch requires a custom token.

Fixes: 7658669 ("multi-arch-test-build: post formal summaries to PR")
Link: https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#pull_request_target
Link: https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/
Signed-off-by: George Sapkin <george@sapk.in>